+++ cvat/apps/engine/serializers.py
@@ -670,7 +670,7 @@ class ShapeSerializer(serializers.Serializer):
     type = serializers.ChoiceField(choices=models.ShapeType.choices())
     occluded = serializers.BooleanField()
     z_order = serializers.IntegerField(default=0)
-    rotation = serializers.IntegerField(default=0, min_value=0)
+    rotation = serializers.FloatField(default=0, min_value=0)
     points = serializers.ListField(
         child=serializers.FloatField(),
         allow_empty=False,
